# output/api-contracts-updated.yaml
openapi: 3.0.0
info:
  title: FormFit Coach API
  description: |
    RESTful API for FormFit Coach platform - connecting fitness enthusiasts with expert trainers for personalized video form analysis.

    ## Authentication
    All authenticated endpoints require either:
    - Session cookie (for web app)
    - JWT token in Authorization header: `Bearer <token>`

    ## Rate Limiting
    Rate limits vary by endpoint and authentication status. Limits are returned in headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Versioning
    API is versioned via URL path (e.g., /api/v1/). Version v1 is current stable version.

  version: 1.0.0
  contact:
    name: FormFit Coach API Support
    email: api-support@formfit.com
servers:
  - url: https://formfit.com/api/v1
    description: Production server
  - url: https://staging.formfit.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
...
    Video:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        fileUrl:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        workoutType:
          type: string
          example: squat
        duration:
          type: integer
          description: Duration in seconds
          nullable: true
        fileSize:
          type: integer
          description: File size in bytes
        status:
          type: string
          enum: [UPLOADING, PROCESSING, READY, FAILED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
    PaginatedResponse:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
paths:
...
  /videos:
    get:
      tags:
        - Videos
      summary: List user's videos
      description: Get paginated list of authenticated user's videos
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [UPLOADING, PROCESSING, READY, FAILED]
        - name: workoutType
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of videos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      videos:
                        type: array
                        items:
                          $ref: '#/components/schemas/Video'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Videos
      summary: Create video record
      description: Create video database record after S3 upload
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - fileUrl
                - workoutType
                - fileSize
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
                fileUrl:
                  type: string
                workoutType:
                  type: string
                fileSize:
                  type: integer
                duration:
                  type: integer
      responses:
        '201':
          description: Video created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        '400':
          description: Validation error or email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...